Add no-plugin-tests for testing Chisel without the compiler plugin
This is a new SBT build unit that symlinks in some files from the normal
chisel project tests, but builds them without the compiler plugin.

(cherry picked from commit 53b620478ddab1faa96512e473fa198f7f1fcf50)
