+++ no-plugin-tests/src/test/scala/chisel3/testers/TestUtils.scala
@@ -1,0 +1,1 @@
+../../../../../../src/test/scala/chisel3/testers/TestUtils.scala
