+++ no-plugin-tests/src/test/scala/chiselTests/ChiselSpec.scala
@@ -1,0 +1,1 @@
+../../../../../src/test/scala/chiselTests/ChiselSpec.scala
